Move generic test fixtures to core pytest plugin and drop legacy PATH alias
Move model_name, time_scale, init_data_handler, add_init_data, and config
fixtures from package-level conftest files into the shared movici_testing
pytest plugin so they are available to all workspace packages without
duplication. Remove the unused PATH alias from the transport-assignment
json_schemas module.
