+++ packages/transport-assignment/tests/conftest.py
@@ -1,15 +1,12 @@
 import dataclasses
-import json
 import shutil
 import typing as t
 import uuid
-from pathlib import Path
 
 import pytest
 from aequilibrae import Project
 
 from movici_simulation_core.core.moment import TimelineInfo
-from movici_simulation_core.model_connector.init_data import DirectoryInitDataHandler
 
 
 @pytest.fixture
@@ -38,60 +35,6 @@ def patch_aequilibrae(monkeypatch, clean_aequilibrae_project):
     if not hasattr(Project, "_new_original"):
         Project._new_original = Project.new
     monkeypatch.setattr(Project, "new", new)
-
-
-@pytest.fixture
-def model_name():
-    return "some_model"
-
-
-@pytest.fixture
-def time_scale():
-    return 1
-
-
-@pytest.fixture
-def init_data_handler(tmp_path_factory):
-    root = tmp_path_factory.mktemp("init_data_handler")
-    return DirectoryInitDataHandler(root)
-
-
-@pytest.fixture
-def add_init_data(init_data_handler):
-    root = init_data_handler.root
-
-    def _add_init_data(name, data: t.Union[dict, str, Path]):
-        if isinstance(data, dict):
-            root.joinpath(f"{name}.json").write_text(json.dumps(data))
-            return
-        path = Path(data)
-        if not path.is_file():
-            raise ValueError(f"{data} is not a valid file")
-        target = (root / name).with_suffix(path.suffix)
-        shutil.copyfile(path, target)
-
-    return _add_init_data
-
-
-@pytest.fixture
-def config(
-    model_config,
-    init_data,
-    time_scale,
-):
-    return {
-        "config": {
-            "version": 4,
-            "simulation_info": {
-                "reference_time": 1_577_833_200,
-                "start_time": 0,
-                "time_scale": time_scale,
-                "duration": 730,
-            },
-            "models": [model_config],
-        },
-        "init_data": init_data,
-    }
 
 
 @pytest.fixture
